input: sensors: add sensors max delay time
Max delay is the delay between two sensor events
corresponding to the lowest frequency that this
sensor supports. When lower frequencies are requested
through batch()/setDelay() the events will be generated
at this frequency instead. It can be used by the
framework or applications to estimate when the batch
FIFO may be full.

Change-Id: Ic8b4bd78061801a48b2774c648758c87b2e20bfc
Signed-off-by: Daqing Chen <chendaqing@codeaurora.org>